leaderboard-implemented-with-redis-sorted-sets

+++ src/main/java/com/sh/roadmap/entity/ScoreEntity.java
@@ -16,7 +16,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @Entity(name = "score")
 public class ScoreEntity extends AbstractBaseEntity {
-    private long score;
+    private Double score;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
